+++ Assets/Scripts/GameLogic/Components/AI.cs
@@ -4,6 +4,15 @@ using Ventura.Util;
 
 namespace Ventura.GameLogic.Components
 {
+    public enum AIType
+    {
+        Null,
+        Static,
+        RandomMovement,
+        Hostile,
+    }
+
+
     public abstract class AI
     {
         protected Actor? _parent;
@@ -19,14 +28,7 @@ namespace Ventura.GameLogic.Components
             this._parent = parent;
         }
 
-        public abstract ActionData? ChooseAction();
-    }
-
-    public enum AIType
-    {
-        Null,
-        Static,
-        RandomMovement,
+        public abstract ActionData? ChooseAction(GameState gameState);
     }
 
 
@@ -42,64 +44,19 @@ namespace Ventura.GameLogic.Components
                     return new StaticAI();
                 case AIType.RandomMovement:
                     return new RandomMovementAI();
+                case AIType.Hostile:
+                    return new HostileAI();
 
                 default:
                     throw new GameException($"Unknown AIType: {DataUtils.EnumToStr(aiType)}");
             }
         }
     }
-    //public class EnemyAI : AI
-    //{
-    //    private List<Vector2Int> _path = new();
-
-    //    public EnemyAI(Actor parent) : base(parent) { }
-
-
-    //    public override GameAction? ChooseAction()
-    //    {
-    //        DebugUtils.Log("EnemyAI.chooseAction");
-
-    //        //var gameState = Orchestrator.Instance.GameState; //FIXME
-
-    //        var target = gameState.Player;
-    //        var gameMap = gameState.CurrMap;
-
-    //        var dx = target.x - _parent.y;
-    //        var dy = target.y - _parent.y;
-
-    //        var distance = Math.Max(Math.Abs(dx), Math.Abs(dy));
-
-
-    //        if (gameMap.Visible[_parent.x, _parent.y])
-    //        {
-    //            //if monster is visible to player, 
-    //            //then player is visible to monster
-    //            if (distance <= 1)
-    //                return new MeleeAction(_parent, dx, dy);
-
-
-    //            _path = GetPathTo(gameMap, target.x, target.y);
-    //        }
-
-    //        if (_path.Count > 0)
-    //        {
-    //            var dest = _path[0];
-    //            _path.RemoveAt(0);
-
-    //            DebugUtils.Log("EnemyAI chose MovementAction");
-    //            return new MovementAction(_parent, dest.x - _parent.x, dest.y - _parent.y);
-    //        }
-
-    //        return new WaitAction(_parent);
-    //    }
-    //}
 
 
     public class StaticAI : AI
     {
-        //public StaticAI(Actor parent) : base(parent) { }
-
-        public override ActionData? ChooseAction()
+        public override ActionData? ChooseAction(GameState gameState)
         {
             return new ActionData(GameActionType.WaitAction);
         }
@@ -108,9 +65,21 @@ namespace Ventura.GameLogic.Components
 
     public class RandomMovementAI : AI
     {
-        //public RandomMovementAI(Actor parent) : base(parent) { }
+        public override ActionData? ChooseAction(GameState gameState)
+        {
+            var actionData = new ActionData(GameActionType.BumpAction);
+            actionData.DeltaPos = RandomUtils.RandomMovement();
+            //FIXME: retry if direction is not walkable
 
-        public override ActionData? ChooseAction()
+            return actionData;
+        }
+
+    }
+
+
+    public class HostileAI : AI
+    {
+        public override ActionData? ChooseAction(GameState gameState)
         {
             var actionData = new ActionData(GameActionType.BumpAction);
             actionData.DeltaPos = RandomUtils.RandomMovement();