+++ Assets/Scripts/GameLogic/Entities/Actor.cs
@@ -7,6 +7,13 @@ namespace Ventura.GameLogic.Entities
     [Serializable]
     public class Actor : Entity, ISerializationCallbackReceiver
     {
+        protected int _currHP;
+        public int CurrHP { get => _currHP; set => _currHP = value; }
+
+        protected int _maxHP;
+        public int MaxHP { get => _maxHP; }
+
+
         [SerializeReference]
         protected Skills _skills = null;
         public Skills Skills { get => _skills; }
@@ -15,6 +22,8 @@ namespace Ventura.GameLogic.Entities
         protected Inventory _inventory = null;
         public Inventory Inventory { get => _inventory; }
         //private Equipment? _equipment;
+
+        public virtual CombatStats? CombatStats { get; }
 
 
         public Actor(string name) : base(name, true) { }
@@ -40,10 +49,6 @@ namespace Ventura.GameLogic.Entities
 
         // -------------------------------------------------
 
-        public override void MoveTo(int x, int y)
-        {
-            base.MoveTo(x, y);
-        }
     }
 
 }